requirements.txt
streamlit>=1.20
sqlalchemy>=1.4
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
openai
sqlparse
pandas
scikit-learn
PyPDF2
